🎨 Consolidate kitchen sink APIs into single API
