+++ examples/query/react/kitchen-sink/src/app/services/posts.ts
@@ -1,5 +1,5 @@
-import { createApi, fetchBaseQuery, retry } from '@reduxjs/toolkit/query/react'
-import { RootState } from '../store'
+import { retry } from '@reduxjs/toolkit/query/react'
+import { api } from './api'
 
 export interface Post {
   id: number
@@ -16,25 +16,7 @@ export interface User {
   phone: string
 }
 
-// Create our baseQuery instance
-const baseQuery = fetchBaseQuery({
-  baseUrl: '/',
-  prepareHeaders: (headers, { getState }) => {
-    // By default, if we have a token in the store, let's use that for authenticated requests
-    const token = (getState() as RootState).auth.token
-    if (token) {
-      headers.set('authentication', `Bearer ${token}`)
-    }
-    return headers
-  },
-})
-
-const baseQueryWithRetry = retry(baseQuery, { maxRetries: 6 })
-
-export const postApi = createApi({
-  reducerPath: 'postsApi', // We only specify this because there are many services. This would not be common in most applications
-  baseQuery: baseQueryWithRetry,
-  tagTypes: ['Posts'],
+export const postsApi = api.injectEndpoints({
   endpoints: (build) => ({
     login: build.mutation<{ token: string; user: User }, any>({
       query: (credentials: any) => ({
@@ -102,8 +84,8 @@ export const {
   useLoginMutation,
   useUpdatePostMutation,
   useGetErrorProneQuery,
-} = postApi
+} = postsApi
 
 export const {
   endpoints: { login, getPost },
-} = postApi
+} = postsApi
