--- examples/query/react/kitchen-sink/src/app/services/split/posts.ts
@@ -1,11 +1,0 @@
-import { emptySplitApi, PostsResponse } from '.'
-
-export const apiWithPosts = emptySplitApi.injectEndpoints({
-  endpoints: (build) => ({
-    getPosts: build.query<PostsResponse, void>({
-      query: () => 'posts',
-      providesTags: (result = []) =>
-        result.map(({ id }) => ({ type: 'Posts', id })),
-    }),
-  }),
-})
