+++ examples/query/react/kitchen-sink/src/app/store.ts
@@ -1,10 +1,7 @@
 import { configureStore, ConfigureStoreOptions } from '@reduxjs/toolkit'
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux'
-import { counterApi } from './services/counter'
-import { postApi } from './services/posts'
-import { timeApi } from './services/times'
+import { api } from './services/api'
 import polling from '../features/polling/pollingSlice'
-import { splitApi } from './services/split'
 import auth from '../features/auth/authSlice'
 
 export const createStore = (
@@ -12,26 +9,18 @@ export const createStore = (
 ) =>
   configureStore({
     reducer: {
-      [counterApi.reducerPath]: counterApi.reducer,
-      [postApi.reducerPath]: postApi.reducer,
-      [timeApi.reducerPath]: timeApi.reducer,
-      [splitApi.reducerPath]: splitApi.reducer,
+      [api.reducerPath]: api.reducer,
       polling,
       auth,
     },
     middleware: (getDefaultMiddleware) =>
-      getDefaultMiddleware().concat(
-        counterApi.middleware,
-        postApi.middleware,
-        timeApi.middleware,
-        splitApi.middleware
-      ),
+      getDefaultMiddleware().concat(api.middleware),
     ...options,
   })
 
 export const store = createStore()
 
 export type AppDispatch = typeof store.dispatch
-export const useAppDispatch = () => useDispatch<AppDispatch>()
+export const useAppDispatch: () => AppDispatch = useDispatch
 export type RootState = ReturnType<typeof store.getState>
 export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector
