+++ examples/query/react/kitchen-sink/src/features/auth/authSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit'
-import { postApi, User } from '../../app/services/posts'
+import { postsApi, User } from '../../app/services/posts'
 import { RootState } from '../../app/store'
 
 const initialState = {
@@ -16,15 +16,15 @@ const slice = createSlice({
   },
   extraReducers: (builder) => {
     builder
-      .addMatcher(postApi.endpoints.login.matchPending, (state, action) => {
+      .addMatcher(postsApi.endpoints.login.matchPending, (state, action) => {
         console.log('pending', action)
       })
-      .addMatcher(postApi.endpoints.login.matchFulfilled, (state, action) => {
+      .addMatcher(postsApi.endpoints.login.matchFulfilled, (state, action) => {
         console.log('fulfilled', action)
         state.user = action.payload.user
         state.token = action.payload.token
       })
-      .addMatcher(postApi.endpoints.login.matchRejected, (state, action) => {
+      .addMatcher(postsApi.endpoints.login.matchRejected, (state, action) => {
         console.log('rejected', action)
       })
   },
